fix(ontology): SJIP-822 fix issues
